[wip] add google and mastodon oauth (#78)
Closes #4 

And thanks to how Google auth works (it doesn't support localhost
domains that aren't just https://localhost) the editor will now load on
`https://localhost`, not `https://editor.com.localhost` anymore.

Also note that `GITHUB_APP_HOST` was unused, so that got removed.

- [x] The setup.js needs to be updated so it can run through the list of
supported auth providers, and ask whether to support that for each one,
and if yes, build the ENV data for that.
- [x] The auth code needs to be able to only set up those auth providers
that are actually ENV-supported
- [x] The docs need information on how to add auth providers.
- [ ] We may need some more tests to cover the new route for adding an
auth provider
- [x] the reservedAccount session object should be stripped on auth
completion.
- [ ] google logins may need the whole "access/refresh" dance, which I
didn't implement in this PR (yet)
- [ ] for mastodon, we obviously have a choice of "which servers to
support" so we may need a follow-up issue for presenting a list of
servers that have an oauth application defined for this platform?
- [x] related, we should probably say which mastodon domain is being
used on the sign up, login, and profile pages.

Locally, I was able to log in with Github, and then on my profile page,
add Google and Mastodon as secondary auth methods, after which I was
able to log in using any of the three providers.

I was also able to delete auth methods, except "the last one".

- [ ] test signup with a new user account (e.g. secondary
github/google/mastodon)