use Playwright to generate project screenshots (#134)
fixes #9 by adding Playwright and using that as a screenshot service for
projects.

This introduces a new env var `USE_LIVE_EMBEDS` that, when true, does
what the platform does right now, but when set to false causes project
screenshots to be taken during project start up, as well as whenever the
project `touch()` functionality gets triggered. These might not be the
best triggers, but they work "well enough" for this to probably be good
enough to land.

Note that in order to set the new var, running `node setup` should ask
whether to use live embeds or not.

Screenshots are written to the `content/__screenshots` dir, using
`projectslug.png` as filename. This dir is exposed on the base domain as
`/screenshots` so that a screenshot for "abc" has a file location
`content/__screenshots/abc.png` and a URL
`https://basedomain/screenshots/abc.png`

The project-card template has been updated to load either a full project
iframe, or screenshot image, and the main page template for anonymous
users has been updated to omit the "in fact, these are real websites!"
paragraph when the live embed var is not set to true.

This should also enable browser based integration testing.